Add files via upload
Tic-Tac-Toe, also known as Noughts and Crosses or Xs and Os, is a classic two-player game played on a 3x3 grid. The game is traditionally played with paper and pencil. The grid consists of nine squares, arranged in three rows and three columns. Players take turns marking a square with their designated symbol (either "X" or "O"). The goal of the game is to create a line of three of one's own symbols horizontally, vertically, or diagonally.i built it with library tkinter python